Ensure consistent log levels in integration tests
Some tests were moved from unit_tests to integration_tests in /a810eb0f8715e182d4e1b1dc1636356b97023711
that depend on log level details. This dependency was not triggered in Github actions workflows
but in komodo bleeding nightly tests.
